Adds file upload to article deposit.
closes #1930
