+++ app/views/works/article_form.html.erb
@@ -24,14 +24,6 @@
 
         <%= render 'contents/edit', content: @content, form_id: @work_form.form_id %>
 
-        <% if Settings.globus.enabled %>
-          <%= turbo_frame_tag 'globus', src: new_content_globus_path(@content), data: { controller: 'dropzone-files', dropzone_files_dropzone_outlet: '.dropzone' }, class: 'dropzone-files' do %>
-            <%= render Elements::SpinnerComponent.new %>
-          <% end %>
-
-          <%= render Elements::HorizontalRuleComponent.new(classes: 'my-3') %>
-        <% end %>
-
         <div id="content-files" class="pane-section">
           <%= turbo_frame_tag @content, 'show', src: content_path(@content), data: { controller: 'dropzone-files' }, class: 'dropzone-files' do %>
             <%= render Elements::SpinnerComponent.new %>
